NGS-general: relocate 'split_fastq.py' to bcftbx.cli.
