+++ bin/split_fastq.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     split_fastq.py: split Fastq by lane
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.split_fastq import main
+if __name__ == "__main__":
+     main()
